Compilation fixes for gcc 3.4.
All of XORP now builds on the prerelease gcc 3.4.  This is actually
quite important, as gcc 3.4 is MUCH closer to ISO compliant than 3.3
(the parser was re-written from scratch).  Hopefully this will make
the Windows port somewhat easier too.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@3537 26435e26-4b25-483a-942d-477236992e84

+++ trunk/xorp/contrib/ospfd/xorp/xrl_target.C
@@ -21,7 +21,7 @@
 // along with this program; if not, write to the Free Software
 // Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA 02111-1307, USA.
 
-#ident "$XORP: xorp/ospfd/xorp/xrl_target.C,v 1.5 2003/05/29 22:37:01 mjh Exp $"
+#ident "$XORP: xorp/ospfd/xorp/xrl_target.C,v 1.6 2003/05/29 23:24:52 mjh Exp $"
 
 #include <time.h>
 #include <vector>
@@ -148,7 +148,7 @@ XrlOspfTarget::common_0_1_get_version(
 				      // Output values, 
 				      string& version)
 {
-    version = c_format("OSPFD v%d.%d", OSPF::vmajor, OSPF::vminor);
+    version = c_format("OSPFD v%d.%d", (int)OSPF::vmajor, (int)OSPF::vminor);
     return XrlCmdError::OKAY();
 }
